refactor: greedily merge integrals.py into pyquantedriver.py
